Decoding ( mostly ) done! * Implemented ( most ) instructions for decoding ) * Cleaned up the avr hpp * Changed instruction num to be 16 bit * Changed opcodes to be 16 bit ( but this will be reverted ) * Renamed registers array * Removed checksum verification as it seems to mess with the Arduino IDE for some reason *

+++ inc/Executable.hpp
@@ -17,7 +17,7 @@
 #include <inttypes.h>
 
 struct Executable {
-    uint8_t instruction_num;    // the internal opcode, mapped to avr.hpp
+    uint16_t instruction_num;    // the internal opcode, mapped to avr.hpp
     uint8_t* operand_1;         // the first operand of an instruction.
                                 // pointer to the register used
     uint8_t* operand_2;         // ditto
